infos basica carteira

+++ carteira.php
@@ -1,3 +1,84 @@
+<?php
+
+require_once('session.php');
+require_once('banco.php');
+
+
+$tipo_usuario = $_SESSION["tipo_usuario"];
+
+if (!array_key_exists("id", $_GET)) {
+  if ($tipo_usuario == 1) {
+    header('location: home.php');
+  }
+  
+  if ($tipo_usuario == 2) {
+    header('location: list-clientes.php');
+  }
+}
+
+$id_carteira = $_GET["id"];
+
+// Se for cliente, verifica se essa carteira percence a ele
+if ($tipo_usuario == 1) {
+  $id_cliente = $_SESSION['id'];
+
+  $queryCarteira = mysqli_query($mysqli,
+    "SELECT * FROM carteira WHERE id='$id_carteira'"
+  );
+
+  if ($queryCarteira && ($result = mysqli_fetch_assoc($queryCarteira)) && (mysqli_num_rows($queryCarteira) > 0)) {
+    if ($result["id_cliente"] != $id_cliente) {
+      header('location: home.php');
+    }
+  } else {
+    header('location: home.php');
+  }
+
+}
+
+// Se for analista, verifica se essa carteira é de algum cliente dele
+if ($tipo_usuario == 2) {
+  $id_analista = $_SESSION['id'];
+  $queryAnalista = mysqli_query($mysqli,
+    "SELECT 
+      id 
+    FROM 
+      usuario 
+    WHERE 
+      id_token in (SELECT id from token WHERE id_analista=$id_analista) AND
+      id in (SELECT id_cliente FROM carteira WHERE id=$id_carteira)"
+  );
+
+  if ($queryAnalista && ($result = mysqli_fetch_assoc($queryAnalista)) && (mysqli_num_rows($queryAnalista) > 0)) {
+    $id_cliente = $result['id'];
+  } else {
+    header('location: list-clientes.php');
+  }
+}
+
+$descricao = '';
+
+$queryCarteira = mysqli_query($mysqli,
+    "SELECT 
+      * 
+    FROM 
+      carteira 
+    WHERE 
+      id = $id_carteira"
+  );
+
+  if ($queryCarteira && ($result = mysqli_fetch_assoc($queryCarteira)) && (mysqli_num_rows($queryCarteira) > 0)) {
+    $descricao = $result['descricao'];
+  } else {
+    if ($tipo_usuario == 1) {
+      header('location: home.php');
+    }
+    
+    if ($tipo_usuario == 2) {
+      header('location: list-clientes.php');
+    }
+  }
+?>
 <!doctype html>
 <html lang="en">
   <head>
@@ -19,17 +100,20 @@
           <div class="col-md-12">
             <div class="h-100 p-5 text-white bg-success bg-gradient rounded-3 row">
               <div class="col-10">
-                <h2>Nome da carteira</h2>
+                <h2><?=$descricao?></h2>
               </div>
-              <div class="col-2 d-flex">
-                <a href="carteira-info-1.php" class="me-3"><button class="w-100 btn btn-lg btn-primary" type="submit">Editar</button></a>
-                <a href="carteira-excluir.php"><button class="w-100 btn btn-lg btn-danger" type="submit">Excluir</button></a>
-              </div>
+              <?php if ($tipo_usuario == 1) { ?>
+                <div class="col-2 d-flex">
+                  <a href="carteira-info-1.php" class="me-3"><button class="w-100 btn btn-lg btn-primary" type="submit">Editar</button></a>
+                  <a href="carteira-excluir.php"><button class="w-100 btn btn-lg btn-danger" type="submit">Excluir</button></a>
+                </div>
+              <?php } ?>
             </div>
         </div>
       </div>
 
       <!-- Investimento, Add operação e histórico de operações -->
+      <?php if ($tipo_usuario == 1) { ?>
       <div class="w-100 row">
         <div class="h-100 p-5 row">
           <div class="col">
@@ -45,6 +129,7 @@
           </div>
         </div>
       </div>
+      <?php } ?>
 
       <!-- Filtro -->
       <h4 class="mb-3 fw-normal">Filtro:</h4>
@@ -73,47 +158,29 @@
           <div class="col"><span>Objetivo (%)</span></div>
           <div class="col"><span>Distância do objetivo</span></div>
         </div>
-        <div class="row mb-3 bg-secondary bg-gradient p-3 rounded text-white">
-          <div class="col"><span>PETR4</span></div>
-          <div class="col"><span>Petróleo</span></div>
-          <div class="col"><span>1200</span></div>
-          <div class="col"><span>R$ 20.20</span></div>
-          <div class="col"><span>R$ 20.000</span></div>
-          <div class="col"><span>15%</span></div>
-          <div class="col"><span>25%</span></div>
-          <div class="col"><span>10%</span></div>
-        </div>
-        <div class="row mb-3 bg-secondary bg-gradient p-3 rounded text-white">
-          <div class="col"><span>PETR4</span></div>
-          <div class="col"><span>Petróleo</span></div>
-          <div class="col"><span>1200</span></div>
-          <div class="col"><span>R$ 20.20</span></div>
-          <div class="col"><span>R$ 20.000</span></div>
-          <div class="col"><span>15%</span></div>
-          <div class="col"><span>25%</span></div>
-          <div class="col"><span>10%</span></div>
-        </div>
-        <div class="row mb-3 bg-secondary bg-gradient p-3 rounded text-white">
-          <div class="col"><span>PETR4</span></div>
-          <div class="col"><span>Petróleo</span></div>
-          <div class="col"><span>1200</span></div>
-          <div class="col"><span>R$ 20.20</span></div>
-          <div class="col"><span>R$ 20.000</span></div>
-          <div class="col"><span>15%</span></div>
-          <div class="col"><span>25%</span></div>
-          <div class="col"><span>10%</span></div>
-        </div>
-        <div class="row mb-3 bg-secondary bg-gradient p-3 rounded text-white">
-          <div class="col"><span>PETR4</span></div>
-          <div class="col"><span>Petróleo</span></div>
-          <div class="col"><span>1200</span></div>
-          <div class="col"><span>R$ 20.20</span></div>
-          <div class="col"><span>R$ 20.000</span></div>
-          <div class="col"><span>15%</span></div>
-          <div class="col"><span>25%</span></div>
-          <div class="col"><span>10%</span></div>
-        </div>
-        
+
+        <?php 
+        $queryAcoes = mysqli_query($mysqli,
+          "SELECT 
+            * 
+          FROM 
+            carteira_acoes
+          WHERE 
+            id_carteira = $id_carteira"
+        );
+      
+        while (($acao = mysqli_fetch_assoc($queryAcoes))) { ?>
+          <div class="row mb-3 bg-secondary bg-gradient p-3 rounded text-white">
+            <div class="col"><span><?=$acao['acao'];?></span></div>
+            <div class="col"><span>XXXXXX</span></div>
+            <div class="col"><span><?=$acao['quantidade'];?></span></div>
+            <div class="col"><span>R$ XXXXXX</span></div>
+            <div class="col"><span>R$ XXXXXX</span></div>
+            <div class="col"><span>XX%</span></div>
+            <div class="col"><span><?=$acao['porcentagem_objetivo'];?>%</span></div>
+            <div class="col"><span>XX%</span></div>
+          </div>
+        <?php } ?>        
      </div>
     </main>
